+++ backend/model/model.go
@@ -380,14 +380,11 @@ type AWSMarketplaceCustomer struct {
 
 type Project struct {
 	Model
-	Name                *string
-	ZapierAccessToken   *string
-	FrontAccessToken    *string
-	FrontRefreshToken   *string
-	FrontTokenExpiresAt *time.Time
-	BillingEmail        *string
-	Secret              *string    `json:"-"`
-	TrialEndDate        *time.Time `json:"trial_end_date"`
+	Name              *string
+	ZapierAccessToken *string
+	BillingEmail      *string
+	Secret            *string    `json:"-"`
+	TrialEndDate      *time.Time `json:"trial_end_date"`
 	// Manual monthly session limit override
 	MonthlySessionLimit *int
 	WorkspaceID         int
@@ -482,6 +479,9 @@ type AllWorkspaceSettings struct {
 	EnableDataDeletion    bool `gorm:"default:true"`
 	EnableNetworkTraces   bool `gorm:"default:true"`
 	EnableUnlistedSharing bool `gorm:"default:true"`
+
+	EnableJiraIntegration  bool `gorm:"default:false"`
+	EnableTeamsIntegration bool `gorm:"default:false"`
 }
 
 type HasSecret interface {