+++ sdk/firstload/CHANGELOG.md
@@ -1,5 +1,17 @@
 # highlight.run
 
+## 9.7.1
+
+### Patch Changes
+
+-   847fdc0: revert single style sheet serialization due to performance regression
+
+## 9.7.0
+
+### Minor Changes
+
+-   ecde630: enable browser OTeL by default
+
 ## 9.5.3
 
 ### Patch Changes