+++ sdk/highlight-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @highlight-run/react
 
+## 8.0.1
+
+### Patch Changes
+
+-   Updated dependencies [847fdc0]
+    -   highlight.run@9.7.1
+
+## 8.0.0
+
+### Patch Changes
+
+-   Updated dependencies [ecde630]
+    -   highlight.run@9.7.0
+
 ## 7.0.3
 
 ### Patch Changes